// src/app/enums/app.schmgmt.enums.ts
export enum SchMgmtEnums {
    BASE_URL = 'http://localhost:8081',
    API_PATH = 'schmgmt/api',
    VERSION = 'v1',
    COURSES_ENDPOINT = 'courses/getAllCourses',
    COURSE_BY_ID_ENDPOINT = 'courses/courseById',
    UPDATE_COURSE_ENDPOINT = 'courses/updateCourse',
    DELETE_COURSE_ENDPOINT = 'courses/deleteCourse',
}